Move component node into separate file
